pytest fixture 'fake_todo' moved to conftest.py
